requirements.txt
colorama==0.4.4
pymongo==4.1.1
python-dotenv==0.20.0
tqdm==4.64.0
